backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "luansilva25", email = "luan250807@gmail.com"},
]
dependencies = [
    "django>=5.0.4",
    "djangorestframework>=3.15.1",
    "django-cors-headers>=4.3.1",
    "djangorestframework-simplejwt>=5.3.1",
    "python-magic>=0.4.27",
    "Pillow>=10.3.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
